+++ pkg/function/udferr/error_test.go
@@ -1,4 +1,4 @@
-package clienttest
+package udferr
 
 import (
 	"fmt"
